Add support for atoms
This adds support for atoms which are literal values in the form: `:foo`. Currently only strings that match the identifier pattern can be used as atoms. Atoms can be used as expressions and also as types. When used as types, their literal values are used for type checking and inference.